+++ resources/views/v1/parts/calendar.blade.php
@@ -1,10 +1,10 @@
-<?php
+@php
 /**
  * Comment: Created to display events in a calendar format
  *          Leverages Laravel Full Calendar package
  * Created: 1/2/2020
  */
-?>
+@endphp
 
 
 <div class="container">
@@ -19,5 +19,6 @@
 </div>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/2.1.3/jquery.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.9.0/moment.min.js"></script>
+<script src="https://unpkg.com/tooltip.js/dist/umd/tooltip.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/fullcalendar/2.2.7/fullcalendar.min.js"></script>
 {!! $calendar->script() !!}
