+++ routes/web.php
@@ -246,10 +246,8 @@ Route::get('/promote/{reg}', 'RegistrationController@promote');
 Route::get('/eventcopy/{slug}', 'EventCopyController@show');
 Route::post('/upload/{folder}/{filetype}', 'AssetController@ajax_store'); // Ajax
 
-// API Routes that circumvent AUTH and mCentric navigation, etc.
-Route::get('/eventlist/{orgID}/{past}/{etID?}', 'EventAPIController@show');
-// The files saved as copies w/2 (EvantAPI, eventlist.blade) need this route
-//Route::get('/eventlist/{orgID}/{past}/{cal}/{etID?}', 'EventAPIController@show');
+// Public API Routes that circumvent mCentric navigation, etc.
+Route::get('/eventlist/{orgID}/{past}/{cal?}/{etID?}', 'EventAPIController@show');
 Route::get('/ticketlist/{eventslug}/{override?}', 'EventController@ticket_listing');
 Route::get('/eventics/{orgID}/{etID?}/{override?}', 'EventController@ics_listing');
